refactor: CORS 설정 config 패키지로 이동
